refactor: extract helpers from AdapterNews
